+++ t/test-genome-DBs/multi/compara/table.sql
@@ -319,7 +319,7 @@ CREATE TABLE `genomic_align_block` (
   `length` int(10) NOT NULL,
   `group_id` bigint(20) unsigned DEFAULT NULL,
   `level_id` tinyint(2) unsigned NOT NULL DEFAULT '0',
-  `direction` int(1) unsigned DEFAULT NULL,
+  `direction` tinyint(1) unsigned DEFAULT NULL,
   PRIMARY KEY (`genomic_align_block_id`),
   KEY `method_link_species_set_id` (`method_link_species_set_id`)
 ) ENGINE=MyISAM AUTO_INCREMENT=6410000010711 DEFAULT CHARSET=latin1;
@@ -725,4 +725,3 @@ CREATE TABLE `synteny_region` (
   PRIMARY KEY (`synteny_region_id`),
   KEY `method_link_species_set_id` (`method_link_species_set_id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1;
-
